+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All Notable changes to `cron` will be documented in this file
 
 - **[BC Break]** `ExpressionField` represents the generic Expression Field abstract class
 - **[BC Break]** `Fields` is the enum representing CRON expression field names.
+- **[BC Break]** `ExpressionAliasError` renamed `AliasError`
+- **[BC Break]** `StartDatePresence` enum renamed `InitialDatePresence`
 
 ### Deprecate